requirements.txt
telethon==1.35.0
python-dotenv==1.0.1
colorama==0.4.6
uuid==1.30
